+++ playground/resolve/__tests__/resolve.spec.ts
@@ -18,7 +18,7 @@ test('deep import with exports field', async () => {
 })
 
 test('deep import with query with exports field', async () => {
-  // since it is imported with `?url` it should return a url
+  // since it is imported with `?url` it should return a URL
   expect(await page.textContent('.exports-deep-query')).toMatch(
     isBuild ? /base64/ : '/exports-path/deep.json'
   )